+++ vps-sdk/src/main/java/com/arvrlab/vps_sdk/di/Module.kt
@@ -1,17 +1,16 @@
 package com.arvrlab.vps_sdk.di
 
 import android.content.Context
+import android.hardware.SensorManager
 import android.location.LocationManager
 import com.arvrlab.vps_sdk.BuildConfig
+import com.arvrlab.vps_sdk.common.CompassManager
 import com.arvrlab.vps_sdk.common.CoordinateConverter
 import com.arvrlab.vps_sdk.data.api.IVpsApiManager
 import com.arvrlab.vps_sdk.data.api.NeuroApi
 import com.arvrlab.vps_sdk.data.api.VpsApiManager
 import com.arvrlab.vps_sdk.data.model.request.RequestVpsModel
-import com.arvrlab.vps_sdk.data.repository.INeuroRepository
-import com.arvrlab.vps_sdk.data.repository.IVpsRepository
-import com.arvrlab.vps_sdk.data.repository.NeuroRepository
-import com.arvrlab.vps_sdk.data.repository.VpsRepository
+import com.arvrlab.vps_sdk.data.repository.*
 import com.arvrlab.vps_sdk.domain.interactor.INeuroInteractor
 import com.arvrlab.vps_sdk.domain.interactor.IVpsInteractor
 import com.arvrlab.vps_sdk.domain.interactor.NeuroInteractor
@@ -61,18 +60,21 @@ internal object Module {
                 .create(NeuroApi::class.java)
         }
         single<INeuroRepository> { NeuroRepository(get(), get()) }
+        single<IPrefsRepository> { PrefsRepository(get()) }
     }
 
     val domain: Module = module {
         factory<INeuroInteractor> { NeuroInteractor(get()) }
-        factory<IVpsInteractor> { VpsInteractor(get(), get()) }
+        factory<IVpsInteractor> { VpsInteractor(get(), get(), get()) }
     }
 
     val presentation: Module = module {
         factory { ArManager() }
         factory { get<Context>().getSystemService(LocationManager::class.java) }
+        factory { get<Context>().getSystemService(SensorManager::class.java) }
+        factory { CompassManager(get()) }
         single { CoordinateConverter() }
-        factory<VpsService> { VpsServiceImpl(get(), get(), get(), get()) }
+        factory<VpsService> { VpsServiceImpl(get(), get(), get(), get(), get()) }
         factory { VpsArViewModel(get(), get(), get()) }
     }