+++ vps-sdk/src/main/java/com/arvrlab/vps_sdk/domain/interactor/INeuroInteractor.kt
@@ -7,6 +7,8 @@ import java.io.IOException
 
 internal interface INeuroInteractor {
 
+    val scaleFactorImage: Float
+
     @WorkerThread
     suspend fun loadNeuroModel(mobileVps: MobileVps)