+++ vps-sdk/src/main/java/com/arvrlab/vps_sdk/domain/model/VpsLocationModel.kt
@@ -4,8 +4,11 @@ import com.arvrlab.vps_sdk.data.LocalizationType
 import com.arvrlab.vps_sdk.data.model.CameraIntrinsics
 
 internal data class VpsLocationModel(
+    val userId: String,
+    val timestamp: Double,
     val locationID: String,
     val gpsLocation: GpsLocationModel?,
+    val compass: CompassModel,
     val nodePose: NodePoseModel,
     val force: Boolean,
     val localizationType: LocalizationType,
@@ -18,8 +21,11 @@ internal data class VpsLocationModel(
 
         other as VpsLocationModel
 
+        if (userId != other.userId) return false
+        if (timestamp != other.timestamp) return false
         if (locationID != other.locationID) return false
         if (gpsLocation != other.gpsLocation) return false
+        if (compass != other.compass) return false
         if (nodePose != other.nodePose) return false
         if (force != other.force) return false
         if (localizationType != other.localizationType) return false
@@ -30,8 +36,11 @@ internal data class VpsLocationModel(
     }
 
     override fun hashCode(): Int {
-        var result = locationID.hashCode()
+        var result = userId.hashCode()
+        result = 31 * result + timestamp.hashCode()
+        result = 31 * result + locationID.hashCode()
         result = 31 * result + (gpsLocation?.hashCode() ?: 0)
+        result = 31 * result + compass.hashCode()
         result = 31 * result + nodePose.hashCode()
         result = 31 * result + force.hashCode()
         result = 31 * result + localizationType.hashCode()