+++ vps-sdk/src/main/java/com/arvrlab/vps_sdk/util/BitmapExtension.kt
@@ -2,6 +2,8 @@ package com.arvrlab.vps_sdk.util
 
 import android.graphics.*
 
+private const val RATIO_16_9 = 16f / 9f
+
 fun Bitmap.toGrayscale(): Bitmap {
     val result = Bitmap.createBitmap(width, height, config)
     val colorMatrix = ColorMatrix()
@@ -11,4 +13,25 @@ fun Bitmap.toGrayscale(): Bitmap {
     Canvas(result)
         .drawBitmap(this, 0f, 0f, paint)
     return result
+}
+
+fun Bitmap.cropTo16x9(): Bitmap {
+    var newWidth: Int = width
+    var newHeight: Int
+
+    newHeight = (width / RATIO_16_9).toInt()
+    if (newHeight == height) return this
+
+    if (newHeight > height) {
+        newHeight = height
+        newWidth = (newHeight * RATIO_16_9).toInt()
+    }
+
+    return Bitmap.createBitmap(
+        this,
+        (width - newWidth) / 2,
+        (height - newHeight) / 2,
+        newWidth,
+        newHeight
+    )
 }