+++ vps-sdk/src/main/java/com/arvrlab/vps_sdk/util/TimestampUtil.kt
@@ -1,0 +1,11 @@
+package com.arvrlab.vps_sdk.util
+
+object TimestampUtil {
+
+    fun getTimestampInSec(): Double =
+        convertMSecToSec(System.currentTimeMillis())
+
+    fun convertMSecToSec(milliseconds: Long): Double =
+        milliseconds / 1_000.0
+
+}