+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v.0.6.1
+February 9, 2022
+
+- VPS service keeps running even when getting HttpException
+- Intrinsics matrix and camera buffer now scales correctly
+- Minor fixes
+
 ## v.0.6.0
 January 18, 2022